Update the sample plugin by using the new transform api
